Refactor date picker logic
